# src/main/resources/static/swagger/productapi.yml
openapi: 3.0.3
info:
  title: 제품 API 명세서
  description: |-
    - 무신사 과제전형 제출용 백엔드 서버의 제품 정보 관련 API 명세서입니다.
  contact:
    email: iceflower01@gmail.com
  version: 0.0.1 SNAPSHOT

servers:
  - url: https://localhost:8080
tags:
  - name: 제품 정보 API
    description: 제품 정보 조회/등록/수정/삭제 기능을 담당하는 API입니다.

paths:
  "/v1/api/product":
    post:
      tags:
        - 제품 정보 API
      summary: 제품 정보 신규등록
      description: 새로운 제품 정보를 DB에 적재한 후, 그 결과를 돌려줍니다.
      operationId: registerProductInfo
      requestBody:
        description: 신규등록할 제품 데이터
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInfoRegisterRequest'
      responses:
        '200':
          description: 작업 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInfo'
        '400':
          description: 신규등록 요청 데이터 포맷이 유효하지 않을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDataValidationErrorResponse'
        '422':
          description: 존재하지 않는 브랜드, 카테고리 정보가 포함되어 있는 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/v1/api/product/list":
    get:
      tags:
        - 제품 정보 API
      summary: 브랜드 정보 리스트 조회
      description: 새로운 브랜드 정보를 DB에 적재한 후, 그 결과를 돌려줍니다.
      operationId: getProductInfoList
      parameters:
        - name: brandId
          in: query
          description: 브랜드 정보 ID
          required: false
          schema:
            type: string
            example: 0JS975W4891XQ
        - name: categoryId
          in: query
          description: 카테고리 정보 ID
          required: false
          schema:
            type: string
            example: 0JS97BHFR9353
        - name: priceFrom
          in: query
          description: 제품가격 조건 from
          required: false
          schema:
            type: integer
            format: int64
            example: 1000
        - name: priceTo
          in: query
          description: 제품가격 조건 To
          required: false
          schema:
            type: integer
            format: int64
            example: 5000
        - name: pageNumber
          description: 0번부터 시작하는 페이지 번호
          in: query
          required: false
          schema:
            type: integer
            format: int32
            example: 0
        - name: pageSize
          in: query
          description: 페이지 사이즈
          required: false
          schema:
            type: integer
            format: int32
            example: 10
      responses:
        '200':
          description: 작업 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProductInfo'
  "/v1/api/product/{productId}":
    get:
      tags:
        - 제품 정보 API
      summary: 개별 제품 정보 조회
      description: 개별 제품 정보를 조회하여 돌려줍니다.
      operationId: getProductInfo
      parameters:
        - name: productId
          in: path
          description: 제품정보 ID
          required: true
          schema:
            type: string
            example: 0JS9VDTS492Z8
      responses:
        '200':
          description: 작업 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInfo'
        '422':
          description: 제품정보를 찾을 수 없을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - 제품 정보 API
      summary: 개별 제품 정보 수정
      description: 개별 제품 정보를 수정한 후 그 결과를 돌려줍니다.
      operationId: correctProductInfo
      parameters:
        - name: productId
          in: path
          description: 제품정보 ID
          required: true
          schema:
            type: string
            example: 0JS9VDTS492Z8
      requestBody:
        description: 수정 요청 데이터
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInfoCorrectRequest'
        required: true
      responses:
        '200':
          description: 작업 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInfo'
        '400':
          description: 수정 요청 데이터 포맷이 유효하지 않을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDataValidationErrorResponse'
        '422':
          description: 제품정보, 브랜드 정보, 카테고리 정보를 찾을 수 없을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 제품 정보 API
      summary: 개별 제품 정보 삭제
      description: 개별 제품 정보를 삭제한 후 그 결과를 돌려줍니다.
      operationId: removeProductInfo
      parameters:
        - name: productId
          in: path
          description: 제품정보 ID
          required: true
          schema:
            type: string
            example: 0JS9VDTS492Z8
      responses:
        '200':
          description: 작업 성공

        '400':
          description: 삭제 요청 데이터 포맷이 유효하지 않을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestDataValidationErrorResponse'
        '422':
          description: 제품정보를 찾을 수 없을 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ProductInfo:
      type: object
      description: 제품정보
      properties:
        productId:
          type: string
          description: 제품ID
          example: 0JS9VDTS492Z8
        productName:
          type: string
          description: 제품명
          example: '[A] 상의'
        categoryId:
          type: string
          description: 카테고리ID
          example: 0JS97BHFR9353
        categoryName:
          type: string
          description: 카테고리명
          example: 상의
        brandId:
          type: string
          description: 브랜드ID
          example: 0JS975W4891XQ
        brandName:
          type: string
          description: 브랜드명
          example: A
        productPrice:
          type: integer
          format: int64
          description: 제품가격
          example: 10100
    PagedProductInfo:
      type: object
      description: 페이징 처리된 제품 정보 리스트
      properties:
        totalPages:
          type: integer
          format: int64
          description: 전체 페이지 갯수
          example: 1
        totalElements:
          type: integer
          format: int64
          description: 전체 엘리먼트 갯수
          example: 9
        size:
          type: integer
          format: int64
          example: 9
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductInfo'
        number:
          type: integer
          format: int32
          description: 현재 페이지 번호
        sort:
          type: object
          description: 정렬 상세
          properties:
            empty:
              type: boolean
              description: 빈 값 여부
              example: false
            sorted:
              type: boolean
              description: 정렬됨 여부
              example: false
            unsorted:
              type: boolean
              description: 정렬안됨 여부
              example: true
        first:
          type: boolean
          description: 첫 페이지 여부
          example: true
        last:
          type: boolean
          description: 마지막 페이지 여부
          example: false
        numberOfElements:
          type: integer
          format: int64
          description: 전체 엘리먼트 갯수
          example: 999919231
        pageable:
          type: object
          description: 페이징 상세
          properties:
            offset:
              type: integer
              format: int64
              description: 오프셋
              example: 9007199254740991
            sort:
              type: object
              description: 정렬 상세
              properties:
                empty:
                  type: boolean
                  description: 빈 값 여부
                  example: false
                sorted:
                  type: boolean
                  description: 정렬됨 여부
                  example: false
                unsorted:
                  type: boolean
                  description: 정렬안됨 여부
                  example: true
            paged:
              type: boolean
              description: 페이지 처리됨 여부
              example: true
            pageSize:
              type: integer
              format: int32
              description: 페이지 사이즈
              example: 10
            pageNumber:
              type: integer
              format: int32
              description: 현재 페이지 번호
              example: 0
            unpaged:
              type: boolean
              description: 페이지 처리 안됨 여부
              example: false
        empty:
          type: boolean
          description: 빈 값 여부
          example: false
    ProductInfoRegisterRequest:
      type: object
      description: 브랜드 신규등록 요청 객체
      properties:
        brandId:
          type: string
          description: 브랜드ID
          example: 0JS975W4891XQ
          required: true
        categoryId:
          type: string
          description: 카테고리ID
          example: 0JS975W4891XQ
          required: true
        productName:
          type: string
          description: 제품명
          example: 신제품 상의
          required: true
        productPrice:
          type: integer
          format: int64
          description: 제품가격
          example: 10000
        requester:
          type: string
          description: 수정 요청자
          example: test@test.com
          required: true
    ProductInfoCorrectRequest:
      type: object
      description: 브랜드 수정 요청 객체
      properties:
        brandId:
          type: string
          description: 브랜드ID
          example: 0JS975W4891XQ
          required: true
        categoryId:
          type: string
          description: 카테고리ID
          example: 0JS975W4891XQ
          required: true
        productName:
          type: string
          description: 제품명
          example: 신제품 상의
          required: true
        productPrice:
          type: integer
          format: int64
          description: 제품가격
          example: 10000
        requester:
          type: string
          description: 수정 요청자
          example: test@test.com
          required: true
    ProductInfoDeleteRequest:
      type: object
      description: 제품 삭제 요청 객체
      properties:
        requester:
          type: string
          description: 삭제요청자
          example: test@test.com
    ErrorResponse:
      type: object
      description: 오류 응답용 객체
      properties:
        code:
          type: string
          description: 오류코드 (영문 대문자와 언더바로 이루어진 문자열)
          example: ERROR_CODE
        message:
          type: string
          description: 오류 메시지 (한국어 문자열)
          example: 이러저러한 오류가 발생하였습니다.
    RequestDataValidationErrorResponse:
      type: object
      description: 요청 데이터 유효성 검사 오류 응답용 객체
      properties:
        code:
          type: string
          description: 오류코드 (영문 대문자와 언더바로 이루어진 문자열)
          example: INVALID_REQUEST_DATA_FORMAT
        message:
          type: string
          description: 오류 메시지 (한국어 문자열)
          example: 요청 데이터 포맷이 유효하지 않습니다.
        detailed:
          type: array
          description: 오류 상세 리스트 (유효하지 않은 필드 및 오류메시지)
          items:
            type: object
            properties:
              value:
                description: 오류가 발생한 필드값
                anyOf:
                  - type: string
                  - type: number
                  - type: integer
                  - type: boolean
                  - type: array
                    items: { }
                  - type: object
              message:
                type: string
                description: 오류 메시지